+++ semcore/d3-chart/src/types/index.d.ts
@@ -38,6 +38,9 @@ export * from './Bar';
 export { default as HorizontalBar } from './HorizontalBar';
 export * from './HorizontalBar';
 
+export { default as CompactHorizontalBar } from './CompactHorizontalBar';
+export * from './CompactHorizontalBar';
+
 export { default as GroupBar } from './GroupBar';
 export * from './GroupBar';