+++ semcore/skeleton/src/index.d.ts
@@ -79,6 +79,7 @@ declare const VennChartSkeleton: typeof Skeleton;
 declare const BubbleChartSkeleton: typeof Skeleton;
 declare const ScatterPlotChartSkeleton: typeof Skeleton;
 declare const RadialTreeChartSkeleton: typeof Skeleton;
+declare const CompactHorizontalBarChartSkeleton: typeof Skeleton;
 
 export default Skeleton;
 export {
@@ -91,4 +92,5 @@ export {
   BubbleChartSkeleton,
   ScatterPlotChartSkeleton,
   RadialTreeChartSkeleton,
+  CompactHorizontalBarChartSkeleton,
 };