+++ semcore/tooltip/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [6.39.1] - 2024-07-19
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/popper` [5.38.0 ~> 5.38.1]).
+
 ## [6.39.0] - 2024-07-17
 
 ### Added